Support loss scale tensor in fp32 when loss grad op produces fp16 output with a new Transform
Summary:
Currently in PopArt, if the training session has an fp16 loss, and an optimizer with loss scaling, then we create an fp16 loss scale tensor in the the IR, and pass it as an input to the loss grad op.

We want to be able to support loss scale values > max(fp16) even in these cases, so it is desirable to support an fp32 loss tensor, and move the downcasting to as late as possible (i.e. inside the loss grad op that consumes the fp32 loss scale tensor and produces anf fp16 gradient tensor).

This is solved here by a new Transform that can be applied after autograd is run. In this transform we:
- identify loss grad ops that take an fp16 loss scale as an input, and can support mixed precision maths, i.e. they produce an fp16 gradient, but can take in an fp32 loss scale. Let's call these 'mixed precision loss grad ops', or MPLGOs
- identify cases in the graph (really only ever 1) of: `lossScale_fp16 -> MPLGO -> grad_fp16`
- replace with: `lossScale_fp32 -> MPLGO -> grad_fp16`
- Turn ops into MPLGOs by modifying their opx grow code only. We start here with NllGradOpx and sgd grad ops that share this same code. More Opxs will be added in a follow-up diff

Fixes T44406

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george

Subscribers: arjunc, oskarb, george

Maniphest Tasks: T44406

Differential Revision: https://phabricator.sourcevertex.net/D51967