Fix incorrect outputSize parameter when creating slice plan and enable planning for gather gradient
Summary:
This diff changes how the slice plan is created so that callers now have to supply an `TensorInfo` object for the output of the operator being planned.  This is used to derive the correct `outputSize` for creating the `popops::embedding::plan`.

This diff also adds slice planning when evaluating the gradient of gather ops.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Maniphest Tasks: T41323, T44085

Differential Revision: https://phabricator.sourcevertex.net/D49995