Make converting of loss scale tesnsor to fp32 optional via a new SessionOption
Summary:
On popart master, the `PreferFp32LossScale` transform always runs during compilation, so always tries to convert the loss scale to fp32, but currently can only do so for a subset of models. If it fails to do this, ir compilation continues silently.

This diff enabling this transform (renamed to `EnsureFp32LossScale` via a SessionOption and having ir compilation fail unless it is successful.

Advantages: if enabled, and ir compilation doesn't error out, the user can now be sure that their model has fp32 loss scale (and therefore there will be no clipping if doing ALS). Disadvantages: the burden of an extra SessionOption.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Subscribers: michaell, romand

Maniphest Tasks: T44406

Differential Revision: https://phabricator.sourcevertex.net/D53216