Remove manual downcasting of 'gs' in the OptimizerDecompose Pattern
Summary:
To prevent underflows of fp16 gradients in the backwards pass (which can happen if doing training on a model with fp16 weights), the user has the option to scale the gradients at the start of the backwards pass (via a 'loss scaling' optimizer paramter), and then downscale them again just before performing the weight update.

Currently, the fp32 downscaling tensor (referred to as `gs` in the OptimizerDecompose Pattern) is downcast to fp16 before being multiplied by the gradient, if the gradient tensor is fp16.

To be able to support loss scaling values of >max(fp16) (see T37578), it is not safe to downcast at this point, as we may be introducing 'accidental underflow'. This means we will be de-scaling the gradients with a different factor to which we scaled them by at the start of the backwards pass.

This diff fixes the issue by removing the manual cast. Since D51035 poplar::multiply can support mixed precision inputs, and casts the output back to fp16 as late as possible, reducing the risk of 'accidental underflow'.

Fixes T47676

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: oskarb

Maniphest Tasks: T47676

Differential Revision: https://phabricator.sourcevertex.net/D53324