Fix loading from a serialized executable when the Ir object passed to popx::serialization::deserializeExecutable has already called its addAdditionalModelProtoTensors method
Summary:
Background
----
The popart Ir contains the state `std::set<Tensor *, PTensorCmp> additionalModelProtoTensors` which contains tensors like optimizer state that need to be saved and loaded to disk when checkpointing.

It is used to modify the Ir's onnx model. This happens when `addAdditionalModelProtoTensors` is called on the Ir.

`additionalModelProtoTensors` and is set (1), and `addAdditionalModelProtoTensors` is called (2) during `Ir::prepare`. When we load from a deserialized model, we require (1) and (2) for checkpointing to work as expected, but we do not necessarily call `Ir::prepare` (the whole point of loading from a cached executable is to avoid long compilation time). Therefore we see (1) and (2) as part of model deserialization.

This diff
----
If we do call `Ir::prepare` on the Ir that we pass to the `deserializeExecutable` method (which is possible using the PopART API), then we call `addAdditionalModelProtoTensors` twice. This is not safe, and can result in an exception. This diff means PopART now supports this case.

It requires that a new boolean flag `bool additionalModelProtoTensorsAdded = false;` is added to the Ir to track if `addAdditionalModelProtoTensors` has been called.

Fixes T38857

Test Plan:
Two new integration tests added:
- `serialize_deserialize_adam_pre_prepared_ir` is a popart reproducer of the poptorch reproducer in T38857
- `serialize_deserialize_adam_pre_prepared_ir_different_optimizer_state` is an additional test that checks that the error-throwing logic added in `deserializeExecutable`. It verifies that an error is thrown if:
  - ir0 is used to create an executable that is then serialized.
  - ir1 is used an input to `deserializeExecutable`
  - ir1 has been prepared, but it is missing an initializer that exists in the serialized executable.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T38857

Differential Revision: https://phabricator.sourcevertex.net/D53903