Force FindPython to always pick virtualenv Python, if there is one
Summary:
Force FindPython to always pick virtualenv Python, if there is one. Previously, it would pick the system Python if it was a newer version.
NOTE: This fix has not been observed to work in all environments

Remove setting of PYBIND11_CPP_STANDARD as the new version of pybind infers this automatically.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, hatemh, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, hatemh, jamesn

Subscribers: jamesn

Maniphest Tasks: T48686

Differential Revision: https://phabricator.sourcevertex.net/D54246