Allow updating the available memory proportion for an operator
Summary:
This diff adds support for updating the available memory proportion through the builder interface.

The implementation will now remove the available memory proportion value when it was previously set.

This is necessary to support PopTorch use cases where a user wants to:

1. provide a global value for the available memory proportion
2. set a fine-tuned value for a particular layer or class of operators.

Test Plan: Added a unittest that tests that the available memory proportion can be now be updated.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: shirazb

Maniphest Tasks: T48675

Differential Revision: https://phabricator.sourcevertex.net/D54267