Add support for FP32 loss scale tensor with FP16 gradients in the general case by adding casting to the input(s) and output(s) of to the first non 'pass-through' op(s) in the backwards pass
Summary:
To support any model that produces an fp16 loss, but the user wants to have the loss scale tensor in fp32, there will be at least one Op in the backwards pass that looks like:
```
fp16_activation -- GradOp -- fp16_gradient
                   /
fp32_gradient ----
```

There are a number of gradient operations that support mixed precision inputs, but this is not generally the case. At least a subset of those are hardcoded in a list in the `EnsureFp32LossScale` transform. If the user enables this transform (i.e. they want to guarantee an fp32 loss scale tensor) but the `GradOp` is not in this list, then the transform will currently abort.

This diff works around this case by inserting `CastOp`s on the fp16 input and fp32 output tensors of `GradOp`. So the above becomes:
```
fp16_act -- CastOp -- fp32_act -- GradOp -- fp32_grad -- CastOp -- fp16_grad
                                   /
fp32_grad ------------------------
```

This diff allows the user to enable the EnsureFp32LossScale transform in their Session with a greater range of models.

Fixes T47331

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand

Subscribers: romand, markk

Maniphest Tasks: T47331

Differential Revision: https://phabricator.sourcevertex.net/D54289