Make subgraph index mapping functions const
Summary: This diff adds const qualifiers to a number of functions implemented by Ops that have subgraphs (If/Loop/Call) that define how in/out indices of Ops map to in/out indices of the subgraph(s). We do this so that we can use these functions in other functions that are const.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, fabiant

Maniphest Tasks: T49816

Differential Revision: https://phabricator.sourcevertex.net/D55271