Prevent recomputation of ops in the final forward PipelineStage along one 'path to the loss' when an op along another path is set to RecomputeType::Checkpoint
Summary:
Currently, if you have a:
- model with >1 path to the loss on the IPU containing the final forward pipeline stage
- and `Pipeline::setRecompute` checkpoints an operation along one path

then inside `Pipeline::setFinalFwdStageRecomputation` the frontier (from which checkpoint annotations are forward-propagated) is non-empty. As a consequence, the recomputation annotations along other paths are unchanged in this functions.

Only if no ops are checkpointed do we initialise the frontier with ops not produced on the IPU.

In this diff,  we start by checkpointing all IpuCopyOps inside `Pipeline::setRecompute`, such that when we reach `Pipeline::setFinalFwdStageRecomputation` the frontier will be initialised with ops at the entry point to all paths to the loss.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oskarb, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oskarb, george

Maniphest Tasks: T49645

Differential Revision: https://phabricator.sourcevertex.net/D55120