Error as early as possible if not able to attach to device
Summary:
It is now no longer possible to call any of `acquire*Device*` or `create*Device` and return a `nullptr`. Instead in these cases an exception is thrown.

This early erroring prevents a more confusing error message when constructing a Session with a nullptr device.

Fixes T46787

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T46787

Differential Revision: https://phabricator.sourcevertex.net/D55471