Add remote buffer handle
Summary:
Added a handle to the remote buffer which encapsulates all the info needed in order to interact with a remote buffer.
In particular, the handle stores the buffer-id, the shape and types of the tensors which will be stored together with how many tensors will be stacked together in the buffer.
The handle also handle logic so that handle objects with the same buffer id contains the same data.

Test Plan:
- Added unitttest for `RemoteBufferHandle`
- Updated unittest for `test_remote_store`

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T48092

Differential Revision: https://phabricator.sourcevertex.net/D55625