Error as early as possible if not able to attach to device if 'allowReturnNullDevice' is false
Summary:
It is now no longer possible to call `create*Device` and return a `nullptr`. Instead in these cases an exception is thrown.

Similarly this behaviour can be turned on for `acquire*Device*` methods by setting `allowReturnNullDevice` to false.
Note, this is default true so as not to change the API. This will be changed to default false in a future release, in order to fix T46787

This early erroring prevents a more confusing error message when constructing a Session with a nullptr device.

Fixes T50242

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T46787, T50242

Differential Revision: https://phabricator.sourcevertex.net/D55471