Clone the logsoftmax input into a different tensor when there are no dim groups.
Summary:
What's changed:
- Copy the "outplace" `logSoftmax` and `logSoftmaxGrad` when there is no dimension grouping.

Rationale:
`logSoftmax` is usually used in a user's `CrossEntropyLoss`. The input to this being their logits.
In some large models this comes from a slice of a serialised matrix multiply. This leads to a very poor mapping and tensor expression. The proposed solution to this is to copy the input to a new tensor, when none of the input dimensions have any detectable grouping.

Resolves T50436

Test Plan: CI tests

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T50436

Differential Revision: https://phabricator.sourcevertex.net/D55830