Add executeOpNTimesEveryMTimes
Summary:
executeOpNTimesEveryMTimes can be applied on popart operation. It effectively modifies how often the operation
is executed regarding the normal iteration steps.
It builds a counter tensor which is incremented by one for each normal iteration step.
There is a simple logic provided by modulo, less and N, M integers, which set up execution flag for if op.
It creates the if op. The op is moved to one branch which is for executing the op.
The second branch "empty subgraph" is meant to be low intensity compute for not executing the op.
For the "empty subgraph" it is possible to connect inputs and outputs
via nop. And it is possible to set up default values of outputs
by creating scalars and expanding them to the outputs.

Example:
When we set N=2, M=6
Execution is:
Yes, Yes, No, No, No, No, Yes, Yes, No, No, No, No, Yes, Yes, No, No, No, No, Yes, Yes

Test Plan: Add test.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Subscribers: evan, #popart

Maniphest Tasks: T47821

Differential Revision: https://phabricator.sourcevertex.net/D54726