Fix scheduling error in MLPerf BERT
Summary: The clipped gradient tensors was not being connected to all ops that required it. This was causing a scheduling error in MLPerf Bert.

Test Plan:
Parameterized one of the gradient clipping test during debugging, and thought it was worth making the change permanent.
Running MLPerf BERT locally now works on my machine, have been unable to create a minimal producer as yet.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, markk, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: fabiant

Maniphest Tasks: T50068

Differential Revision: https://phabricator.sourcevertex.net/D56434