Use slice plan in SparseAccumulateOpx
Summary:
Create and use a slice plan for the `multiUpdateAdd` in SparseAccumulateOpx. This should in general have very little impact on memory usage, but will target a more efficient vertex for the operation, and spread it to more tiles, resulting in a significant reduction in the total number of cycles needed to perform this operation.
Fixes T51966.

Test Plan: `./test.sh popart`

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, fabiant

Subscribers: matthewha

Maniphest Tasks: T51966

Differential Revision: https://phabricator.sourcevertex.net/D57549