Add pytest name to deviceaccess log name.
Summary: The device access log is a log that is optionally enabled in cmake to see an audit log of tests attaching/detaching from devices. There is a validation script in the works (D57970) that checks the events in this log follow a certain order, and tests do not contribute to the possibility of deadlocking on device-attachment. The test name used in the device log is the name used in ctests. However, as `pytest` is ran with the `--forked` option, many python test functions run in parallel making it seem like some tests acquire multiple devices before detaching. To fix this, the test name needs to be more finegrained -- we add the actual pytest name here.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Subscribers: evan

Maniphest Tasks: T51433

Differential Revision: https://phabricator.sourcevertex.net/D58292