Make outliner and other transforms respect replica-equalness
Summary:
Depends on D48465.

We currently use `Op::getSubgraphEquivId` in a number of transforms (particularly the outliner) to determine if two Ops are equivalent. That is, the outliner's condition as to whether some `op0` and `op1` can be outlined is essentially `op0->getSubgraphEquivId() == op1->getSubgraphEquivId()`. This function returns a string that contains, e.g., the Op's attribute values, the shapes of input/output tensors, and so on. The problem with is as follows: when we use the `ReplicaEqualAnalysis` to work out for every Op whether it's inputs/outputs are replica equal, as we can only have one such judgement on replica-equalness for each Op and can't distinguish per call site, we do not want a situation where an Op in a subgraph is called by different call sites that do not agree on replica-equalness. For example, an `IdentityOp` somewhere early in the fwd pass might not have replica-equal inputs/outputs, but may match shape with an `IdentityOp` that follows a replicated all-reduce, so whose inputs/outputs *are* replica-equal. If the two match on all other attributes then the current implementation may outline these `IdentityOps` together and cause problems during `ReplicaEqualAnalysis`. The changes in this diff stop this from happening.

Changes:
* `Op::getSubgraphEquivId` -- To facilitate incorporating replica-equal information into `getSubgraphEquivId` this diff adds a parameter to this function that enables adding an external state to be taken into account in the equivalence checks, by means of a mapping from string keys to `popart::any` values.
* `ReplicaEqualAnalysis` -- Expanded to have a function `getOpAttrs` that turns replica-equal info for an Op into a mapping that can be used as a parameter to the new `Op::getSubgrapheEquivId`.
* `MergeLoops`/`SubgraphOutline`/`MergeDuplicateOps`/`BatchSerialisation` transforms -- Adapted to 1) do a `ReplicaEqualAnalysis` and 2) call the new `getSubgraphEquivId`. Note that 2) in places was awkward because there was a lot of templated code that can run on arbitrary types and assumes a function `getSubgraphEquivId()` exists on that type without parameters. A different type, `Blip` with this property is often used in unit tests. The way I've gone around fixing this is by leaving the subgraph outline code untouched, and by wrapping Ops in WrappedOps to provide the expected interface. NOTE: This is an alternative to D56790 and was chosen due to review comments.
* Because the `ReplicaEqualAnalysis` is now called in more places, and I initially do not want this to have an impact in the default code path (this is initially an opt-in feature, until we are happy it works) I have guarded some logic behind the `opts._enableRngStateManagement` option. This is so this code won't run unless the user opts in. This will be removed with T48752.
* I've added some better `operator<<` support for `popart::any` because that was useful for the unit tests.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: fabiant

Maniphest Tasks: T48510

Differential Revision: https://phabricator.sourcevertex.net/D57988