Made the first step to integration popef metadata with serialization path in popart
Summary:
Partially addressed Ref T43977
The change is relied on adding popef metadata and popef tensors to a serialized executable.
Note that currently tensor data and other stuff are duplicated, because those data are located in two places.
Therefore, in the next step popef metadata and tensors have to be integrated with popart metadata.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, anthonyb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, anthonyb

Subscribers: fabiant, anthonyb, jamesn, matthewha

Maniphest Tasks: T43977

Differential Revision: https://phabricator.sourcevertex.net/D58976