Add an overload of Builder::setAvailableMemoryProportion for multi-output nodes
Summary:
The existing prototype for setAvailableMemoryProportion only takes one tensor as an input, so can't
be used to set the available memory proportion for multi-output nodes like LSTMs. This change adds
an overload that takes multiple tensor inputs and looks for a node that outputs all of them.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, fabiant

Maniphest Tasks: T48680

Differential Revision: https://phabricator.sourcevertex.net/D58729