Fix unwinding issue in MaxOpx
Summary:
This diff includes a test that replicates the issue seen in T53871.
Without the included fix, the test failed with the error:
```
E       popart_core.popart_exception: Regions are of different rank (1 vs. 2) in sub
```
The cause of this issue is that `MaxOpx` and `MinOpx` inherited `ElementWiseUnaryOpx` despite having multiple inputs. It looks like this has been an issue before, D37881 overrode the inherited unwinding code for `MinOpx`.
I've taken the fix for `MinOpx` from D37881 and applied it to `MaxOpx`, and that causes the reproducer test to pass.
This code should probably be factored out into an `ElementWiseNaryOpx` or something, although I don't know if it's really worth it.
The other option would be to make `MaxOpx` a binary op, and have a pattern to replace instances of `max(a, b, c)` with `max(c, max(a, b))`.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T53871

Differential Revision: https://phabricator.sourcevertex.net/D59260