Added additional API for loading executables from stream
Summary:
Reader created in loadExecutableFromStream was unsafe because it took a reference to stream instead of shared ptr.
More details here: Ref T53405

This change contains an additional API for loadExecutableFromStream with shared_ptr as an argument.
The same change was applied to the internal class - Reader.
Besides there are changes for executable_serialization_test.cpp:
	1) Added test for checking load/save functionality (cache functionality was tested only till now).
	2) Reader is created through previous vulnerability.
	3) Small refactor regarding creation files and directories for test purposes.
	4) Test directory is removed after finishing work. Currently these tests create a lot of trash directories.

To fully address the issue reported in the task additional changes are expected:
	1) Poptorch should use the new loadExecutableFromStream API.
	2) Old vulnerable API needs to be deleted.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, markk

Subscribers: fabiant

Maniphest Tasks: T53405

Differential Revision: https://phabricator.sourcevertex.net/D59354