Shortened popart.ir.op docstring summaries
Summary: According to PEP 257 the summary of a docstring is separated from the rest of the docstring by a blank line. I am using this definition to extract docstring summaries automatically and some of them are a bit verbose to put in a table, so this makes for some terser docstring summaries.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Subscribers: michaell, yanz

Maniphest Tasks: T47521

Differential Revision: https://phabricator.sourcevertex.net/D59475