Make __all__ explicit, avoid wildcard import (in popart.ir).
Summary:
This diff changes two things. We set `__all__` explicitly, instead of relying on `dir()`. This is because `dir()` is tainted and is not limited to just what is reported (see https://docs.python.org/3/reference/import.html#submodules). Secondly, we avoid doing wildcard imports. This is to make it clear what objects are imported from where and avoid namespace pollution. It is my interpretation of the Python best practices that this is what we should do. This is also generally what other packages do:

  - https://github.com/pandas-dev/pandas/blob/main/pandas/__init__.py
  - https://github.com/numpy/numpy/blob/main/numpy/__init__.py
  - https://github.com/scikit-learn/scikit-learn/blob/main/sklearn/__init__.py

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, joshl, michaell

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, joshl, michaell

Subscribers: michaell, joshl

Maniphest Tasks: T47521

Differential Revision: https://phabricator.sourcevertex.net/D59769