Add popart.ir pipelining test
Summary:
This diff adds a test for demonstrating pipelining in popart.ir.
In order to focus on the popart.ir model execution of the test has been abstracted to the `PipelineTestRunner` class.
The test is compared with a reference model written in `torch` which has been separated to its own file.

Test Plan:
- Add `test_0_serial_inference.py` (modified from `test_fwd_pipeline.py`) to demonstrate the model of interest without pipelining.
- Add `test_1_pipelined_inference.py` which demonstrates the model of interest with pipelining.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, fabiant, markk, oskarb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: yanz

Maniphest Tasks: T37127

Differential Revision: https://phabricator.sourcevertex.net/D53488