Factor out core main graph autodiff logic into separate class
Summary:
This refactors the main graph autodiff logic into a separate class.

It removes a lot of the lamdas, which as noted in a comment below may be less desired. Given the cyclomatic complexity of the autodiff function was probably quite high I chose to go without lambdas.

Creates a new autodiffer  class that does the main graph autodiff.

DOES NOT fix the issue where non-differentiable ops were not correctly handled whilst auto-diffing. These issue presents itself in the child revision D58629 (at time to writing, to be fixed). This is a pure refactor.

Test Plan: Adds a test for the upcoming child revision D58629 to check for diffing on subgraphs where not all grads are requested.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Subscribers: markk

Maniphest Tasks: T47692

Differential Revision: https://phabricator.sourcevertex.net/D54551