Fix 'ExplicitRecomputation_Case' test and re-enable
Summary:
As part of fixing compiler warnings in D58585, this test was found to be disabled, with no TODO/task number. I have not investigated why this test was disabled. This diff re-enables the test to fill the coverage hole for the `ExplicitPipeline` transform.

The re-enabled test on the following check:
```
        // All Forward pass Ops must have recomputeType set to Recompute
        // if explicit recomputation is turned off and path is not from loss
```
This is because it defines a 'forward pass op' here as one where `op->fromLoss == PathFromLoss::No`. However, if we instead define a 'forward pass op' as one where `op->fromLoss == PathToLoss::Yes` (which I believe makes more sense in this context), the the test's check passes. It's not clear to me why the test previously used this other condition.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell, zigmasb, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell, fabiant

Maniphest Tasks: T53563

Differential Revision: https://phabricator.sourcevertex.net/D59905