Inference Workaround patch when using VariableSettings
Summary: As using VariableSettings changed how VariableTensors are created (and their potential buffer size), the optimization of using `consts` during inference would throw an error if non-standard VariableSettings were used. This diff makes it so that VariableSettings incompatible with this optimization will ignore the optimization for the incompatible variable.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T54483

Differential Revision: https://phabricator.sourcevertex.net/D60582