Inplacing ambiguity detection for popart.ir
Summary:
This diff adds inplacing ambiguity detection. E.g. if a user adds a a sequence of inplacing ops that would lead to ambiguous behaviour. Such as:
```
   *
   * a <- variable({})
   * b <- modify(a)
   * c <- modify(a),
   *
   *    variable
   *       |
   *    +--+-----+
   *    |        |
   * modify    modify.
   *
   * without any topological constraints between b and c, the final value of
   * 'a' is potentially ambiguous. Suppose that b is 'sqrt' and c is 'sin'.
   * The final value of 'a' depends on the schedule:
   *
   * (a,b,c) : a <- sin(sqrt(a))
   * (a,c,b) : a <- sqrt(sin(a)).
```
i.e. the order is unconstrained and can effect the order of operations and can produce different results.

(Example taken from poprithms docs)

See D56051 for further information about what constitutes an ambiguity. This  is pretty much just a straight implementation of poprithm's ambiguity detection, with some methods to give popART specific info. Also added a couple of popart.ir inplace operators for testing.

I removed the growAliasModel code from host store, as it was attempting to use the host tensor in the alias model. On talking to @jamesn I think this is difficult to represent in a alias model, and it is at worst an edge case that this host tensor would have an inplace ambiguity on it.

resolves T53736

Test Plan: Test added.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, jamesn, oskarb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, jamesn, oskarb

Subscribers: oskarb, michaell, markk, jamesn

Maniphest Tasks: T53736

Differential Revision: https://phabricator.sourcevertex.net/D59190