Changed documentation regarding gradsRequired and added test
Summary:
The autodiff implementation already guarantees that the expected outputs of a grad graph follow exactly the required gradient list, when given. However, we didn't test this, and the doxygen comments of the `autodiff.hpp` and `autodiff.py` files didn't really give this as a guarantee. As it turns out, this property is useful for `popart.ir` users, so we now make it a formal guarantee with this diff.

Changes are:

* Adapt documentation in `autodiff.py` and `autodiff.cpp` to add the guarantee on expected outputs.
* Add c++ tests to extensively check this property. To do this, I added a new test graph to the test graph test util.
* Adapt existing autodiff python test to check this.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: shirazb

Maniphest Tasks: T48894, T48893

Differential Revision: https://phabricator.sourcevertex.net/D61270