Refactor binary elementwise opxs for new snap API
Summary: Snap has added maybeInplace functions in the snap::popops api with the intention to remove some of the conditional inplacing in popart, specifically the conditional inplacing that queries the tile imbalance. The check for `isParallellWriteable` is being left in popart as it will be used as an early test of popirs alias analysis.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, fabiant, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: jamesn

Maniphest Tasks: T55117

Differential Revision: https://phabricator.sourcevertex.net/D61304