Add 'GradientTensorTrackingMethod' setting to 'AutomaticLossScalingSettings' to allow the user to influence the selection of gradient tensors whose statistics are tracked in the AutomaticLossScaling transform
Summary:
The `automaticLossScalingSettings` SessionOption is used when constructing a TrainingSession to configure behaviour of the AutomaticLossScaling transform.

This diff adds a new attribute to this struct, `GradientTensorTrackingMethod`, which allows the user to influence the tensors whose statistics are tracked in order to decide how to update the loss scale factor. The user can choose from the existing two options, `ConvAndMatmulGradients`, and `GradientsOfUserSpecifiedTensors`.

Additionally, a new gradient tensor tracking method, `AllNonViewChangingGradientTensors`, has been added. This has been also been set as the new default, as requested by Apps. Since this is marked as an experimental feature in sessionoptions.hpp docs, this is safe to do without waiting for a major revision change.

Fixes T52592

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand, michaell

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand

Subscribers: sergiop, arjunc

Maniphest Tasks: T52592

Differential Revision: https://phabricator.sourcevertex.net/D61494